Hide backoff on stop supervisor implementation #26156
